whos_who/requirements.txt
pandas==1.3.4
opencv-python==4.5.4.60
Pillow==8.4.0
fastparquet==0.7.2